+++ Sources/PulseUI/Views/RichTextView/WrappedTextView.swift
@@ -99,6 +99,7 @@ struct WrappedTextView: NSViewRepresentable {
         }
 
         textView.delegate = context.coordinator
+        textView.isEditable = viewModel.isEditable
 
         textView.attributedText = viewModel.originalText